% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\encoding{UTF-8}
\name{average_color}
\alias{average_color}
\title{average_color}
\usage{
average_color(img)
}
\arguments{
\item{img}{the image (a array)}
}
\description{
return the average color from an image
}

