% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\encoding{UTF-8}
\name{draw}
\alias{draw}
\title{draw}
\usage{
draw(folder)
}
\arguments{
\item{folder}{the path of the folder}
}
\description{
transform a folder of scrennshot into a synthetical picture
}
\examples{
\dontrun{
#options(ffmpeg = "C:/ffmpeg-3.1.4-win64-static/bin/ffmpeg.exe") # windows user
library(magrittr)
gen_screenshot(movie = "D:/mymovie.avi") \%>\%
draw()
}
}

