% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\encoding{UTF-8}
\name{folder_average_color}
\alias{folder_average_color}
\title{folder_average_color}
\usage{
folder_average_color(folder)
}
\arguments{
\item{folder}{the path of the folder}
}
\description{
return the average color from all folder's images
}

