% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\encoding{UTF-8}
\name{gen_screenshot}
\alias{gen_screenshot}
\title{gen_screenshot}
\usage{
gen_screenshot(movie, folder = tempfile(), every = 30, stdout = "")
}
\arguments{
\item{movie}{the path of the movie}

\item{folder}{the path of the folder}

\item{every}{timbe between each screenshot in seconds}
}
\description{
transform a movie into lot of screenshot
}
\examples{
\dontrun{
gen_screenshot(movie = "D:/mymovie.avi") \%>\%
browseURL()
}
}

