% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spongecake.R
\docType{package}
\name{spongecake}
\alias{spongecake}
\alias{spongecake-package}
\title{spongecake}
\description{
this package transform a movie into a simple image...
}
\examples{
\dontrun{
#options(ffmpeg = "C:/ffmpeg-3.1.4-win64-static/bin/ffmpeg.exe") # windows user
library(magrittr)
gen_screenshot(movie = "D:/mymovie.avi") \%>\%
draw()
}
}

